Add files via upload
Hi guys,
I added a timer and took out the clock. I'm trying to get it so when the ball hits the left wall, the timer resets back to 15 seconds and starts counting down immediately. If the ball doesn't hit the paddle and after that the wall before the 15 seconds is up, it says "Game Over" and "Click to Restart" exactly like how it does if the ball doesn't hit the paddle and goes out. I am struggling with getting the timer to reset to exactly 15 seconds and continue from there, rather than continuing back at the time the clock was at prior to hitting the wall. I also can't get the game to reset when clicked, like I had been able to before trying to add the timer. Any ideas on how to fix it?